# 01-access-data-compute/provenance-server/openapi.yaml
openapi: "3.1.0"
info:
  title: GA4GH Provenance Server
  version: "0.1.0"
  description: |
    A Beacon-style API for querying workflow provenance and data elements across
    GA4GH federated sites.

    Each data element can be linked to the workflow (registered via Dockstore / TRS)
    that produced it, enabling cross-site provenance queries similar to the Beacon
    Network for variant data.

    ## Key concepts

    - **ProvenanceRecord** — links a data element (DRS object) to a specific
      workflow version and execution (WES run / TES task).
    - **Workflow** — registered using a canonical TRS URI
      (e.g. `trs://dockstore.org/workflow/...`), optionally auto-resolved from
      the Dockstore API.
    - **Federated query** — fans out to configured peer nodes and aggregates
      results (Beacon Network pattern).
    - **RO-Crate export** — any query result can be downloaded as a standards-
      compliant RO-Crate zip archive.

  contact:
    name: GA4GH Federated ML Hackathon
    url: https://github.com/kellrott/ga4gh-fedml-hackathon-connect2026
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /provenance/query:
    post:
      tags: [query]
      operationId: queryRecords
      summary: Query provenance records (Beacon-style)
      description: |
        Filter provenance records by workflow, version, site, or data element.
        Mirrors the Beacon v2 query/response pattern.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvenanceQuery"
      responses:
        "200":
          description: Query results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeaconQueryResponse"
  /provenance/query/export:
    post:
      tags: [query]
      operationId: exportRecords
      summary: Export matching records as an RO-Crate zip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvenanceQuery"
      responses:
        "200":
          description: RO-Crate zip archive
          content:
            application/zip:
              schema:
                type: string
                format: binary
  /federated/query:
    post:
      tags: [federation]
      operationId: federatedQuery
      summary: Fan-out query to all peer nodes
      description: |
        Queries this node's peers in parallel and aggregates results.
        Peer nodes are configured via the PROVENANCE_PEER_NODES environment variable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvenanceQuery"
      responses:
        "200":
          description: Aggregated federated results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeaconFederatedResponse"
...
components:
...
  schemas:
    BeaconMeta:
      type: object
      properties:
        beacon_id:
          type: string
        api_version:
          type: string
          default: "v0.1.0"
        returned_granularity:
          type: string
          default: "record"
...
    ProvenanceRecordCreate:
      type: object
      required: [data_element_id, workflow_id]
      properties:
        data_element_id:
          type: string
          description: DRS object ID or other resolvable data element identifier
          example: "drs://drs.example.org/abc123"
        workflow_id:
          type: string
          description: TRS URI of the producing workflow
          example: "trs://dockstore.org/workflow/github.com/broadinstitute/gatk-workflows/broad-prod-wgs-germline-snps-indels"
        workflow_version:
          type: string
          nullable: true
        execution_id:
          type: string
          nullable: true
          description: WES run ID or TES task ID
        site:
          type: string
          nullable: true
        parameters:
          type: object
          nullable: true
          additionalProperties: true
        execution_timestamp:
          type: string
          format: date-time
          nullable: true
        extra_metadata:
          type: object
          nullable: true
          additionalProperties: true
    ProvenanceRecordRead:
      allOf:
        - $ref: "#/components/schemas/ProvenanceRecordCreate"
        - type: object
          properties:
            id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
    ProvenanceQuery:
      type: object
      properties:
        workflow_id:
          type: string
          nullable: true
        workflow_version:
          type: string
          nullable: true
        site:
          type: string
          nullable: true
        data_element_id:
          type: string
          nullable: true
        include_parameters:
          type: boolean
          default: false
        limit:
          type: integer
          default: 100
        offset:
          type: integer
          default: 0
...
    BeaconQueryResponse:
      type: object
      properties:
        meta:
          $ref: "#/components/schemas/BeaconMeta"
        response:
          type: object
          properties:
            numTotalResults:
              type: integer
            results:
              type: array
              items:
                $ref: "#/components/schemas/ProvenanceRecordRead"
    BeaconFederatedResponse:
      type: object
      properties:
        meta:
          $ref: "#/components/schemas/BeaconMeta"
        response:
          type: object
          properties:
            numTotalResults:
              type: integer
            results:
              type: array
              items:
                type: object
                additionalProperties: true
            errors:
              type: array
              items:
                type: string
            queriedNodes:
              type: array
              items:
                type: string
...